% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaucat_dat.R
\docType{data}
\name{blaucat_dat}
\alias{blaucat_dat}
\title{BlauCat Dataset}
\format{
An object of class \code{data.frame} with 67 rows and 113 columns.
}
\usage{
blaucat_dat
}
\description{
BlauCat Dataset
}
\keyword{datasets}
